Able to send data to a database
First and foremost, I've created the extremely basic-looking UI for the Login and Registration System, and only the registration system is semi-functional. You're able to put information in the input boxes, and hit register. The data successfully makes it's way to a database. The page can successfully connect and communicate with a database.