Reworked the timeLogs table to have four entries
Had to rework the timeLogs table to have four entries, since if the user was clocked in and the day passes, the timeSpent entry will break, and go to a strange negative value. To get around this, I converted clockInTime to clockIn, and clockOutTime to clockOut, which includes both the date AND time. The math that calculates the timeSpent accounts for both date and time, which is then able to calculate the correct value.